+++ AnalyzerTest/MathTests.cpp
@@ -17,15 +17,27 @@ namespace AnalyzerTest
 		TEST_METHOD(Math_RiseAndFall)
 		{
 			using namespace DirectX;
-			XMVECTOR vPrevious = XMVectorSet(1, 1, 2, 2);
-			XMVECTOR vCurrent = XMVectorSet(1, 2, 1, 2);
+			XMVECTOR vPrevious = XMVectorSet(1, 1, 2, 1);
+			XMVECTOR vCurrent = XMVectorSet(1, 2, 1, 11);
 			XMVECTOR vResult = XMVectorReplicate(std::numeric_limits<float>::max());
 			AudioMath::ApplyRiseAndFall(&vPrevious, &vCurrent, &vResult, 1, 1, 3);
 
 			Assert::AreEqual(1, vResult.m128_f32[0], 0.001f, L"1->1");
 			Assert::AreEqual(1.63212055f, vResult.m128_f32[1], 0.001f, L"1->2");
 			Assert::AreEqual(1.04978707f, vResult.m128_f32[2], 0.001f, L"2->1");
-			Assert::AreEqual(2, vResult.m128_f32[3], 0.001f, L"2->2");
+			Assert::AreEqual(7.3212055f, vResult.m128_f32[3], 0.001f, L"1->11");
+		}
+		TEST_METHOD(Math_ConvertToLog)
+		{
+			using namespace DirectX;
+			XMVECTOR vInput[2] = { XMVectorSet(-1.0f,0.0f,1e-6f,0.1f),XMVectorSet(1.0f,10.0f,std::numeric_limits<float>::max(),INFINITY) };
+			XMVECTOR vResult[2];
+			AudioMath::ConvertToLogarithmic(vInput, vResult, 2,-100.0f,20.0f);
+			float expected[] = { -100,-100,-100,-20,0,20,20,20 };
+			for (size_t i = 0; i < 8; i++)
+			{
+				Assert::AreEqual(expected[i], ((float*)vResult)[i]);
+			}
 		}
 	};
 }
