+++ AudioMath/AudioAnalyzer.h
@@ -2,12 +2,15 @@
 #include <DirectXMath.h>
 #include "AudioBuffer.h"
 #include <memory>
+#include <mutex>
+
+
+using namespace DirectX;
 
 namespace AudioMath
 {
 	class CAudioAnalyzer
 	{
-		UINT32 _inputSampleRate;
 		UINT32 _inputChannels;
 		UINT32 _fftLength;
 		UINT32 _fftLengthLog2;
@@ -21,6 +24,7 @@ namespace AudioMath
 		DirectX::XMVECTOR *_pFftBuffers;
 		float _fFftScale;	// 2/N scale factor for fft output
 		std::shared_ptr<CAudioBuffer> _spInputBuffer;
+		std::mutex _inputBufferAccess;
 
 		void AllocateBuffers();
 		void FreeBuffers();
@@ -28,9 +32,8 @@ namespace AudioMath
 		CAudioAnalyzer(size_t inputBufferSize);
 		~CAudioAnalyzer();
 
-		void ConfigureInput(UINT32 inputSampleRate, UINT32 inputChannels)
+		void ConfigureInput(UINT32 inputChannels)
 		{
-			_inputSampleRate = inputSampleRate;
 			_inputChannels = inputChannels;
 			_spInputBuffer->SetFrameSize(_inputChannels);
 		}
@@ -39,14 +42,16 @@ namespace AudioMath
 		// Returns true if output data is available
 		bool IsOutputAvailable();
 		
-		// Adds data to the input buffer
-		void AddInput(float *pData, size_t frameCount, long frameIndex = -1);
+		// Adds data to the input buffer. NB size is in samples!
+		void AddInput(float *pData, size_t sampleCount, long frameIndex = -1);
 		
 		// Fetches next set of data from buffer and analyzes
-		bool Step(long *pPosition,float *pRMS, float *pPeak, DirectX::XMVECTOR *pSpectrum);
+		bool Step(long *pPosition, DirectX::XMVECTOR *pRMS, DirectX::XMVECTOR *pPeak, DirectX::XMVECTOR *pSpectrum);
 		
 		// Flushes all data from the input buffer
 		void Flush();
+
+		long GetPosition() const { return _spInputBuffer->GetPosition(); }
 	};
 }
