+++ AudioVisualizer/AudioVisualizer.idl
@@ -105,7 +105,6 @@ namespace AudioVisualizer
 	[version(1.0)]
 	interface IVisualizationData : IInspectable
 	{
-		/*[propget] HRESULT Dimensions([out] UINT32* lsize, [out, retval, size_is(, *lsize)] UINT32**value);*/
 		[propget] HRESULT AmplitudeScale([out,retval] AudioVisualizer.ScaleType *scale);
 	}
 
@@ -125,8 +124,10 @@ namespace AudioVisualizer
 	[uuid(FA6EDA25-96DC-473E-AA4C-5F67C2307134),version(1.0)]
 	interface IScalarData : IInspectable
 	{
-		HRESULT ConvertToLogAmplitude([in] float minValue, [in] float maxValue, [out, retval] VectorData **result);
-		HRESULT ApplyRiseAndFall([in] VectorData *previous, [in] Windows.Foundation.TimeSpan riseTime, [in] Windows.Foundation.TimeSpan fallTime, [in] Windows.Foundation.TimeSpan timeFromPrevious, [out, retval] VectorData **result);
+		HRESULT ConvertToLogAmplitude([in] float minValue, [in] float maxValue, [out, retval] ScalarData **result);
+		HRESULT ApplyRiseAndFall([in] ScalarData *previous, [in] Windows.Foundation.TimeSpan riseTime, [in] Windows.Foundation.TimeSpan fallTime, [in] Windows.Foundation.TimeSpan timeFromPrevious, [out, retval] ScalarData **result);
+		[propget] HRESULT Size([out, retval] UINT32 *size); 
+		[propget] HRESULT Values([out] UINT32* lsize, [out, retval, size_is(, *lsize)] float **data);
 	};
 
 	[version(1.0)]
@@ -134,8 +135,8 @@ namespace AudioVisualizer
 	{
 		[default] interface IVisualizationData;
 		interface IScalarData;
-		interface Windows.Foundation.Collections.IVectorView<float>;
 	}
+
 	[version(1.0)]
 	runtimeclass VectorData
 	{
@@ -164,8 +165,9 @@ namespace AudioVisualizer
 	[version(1.0)]
 	interface IVisualizationSourceStatics : IInspectable
 	{
-		HRESULT CreateSourceFromMediaElementAsync([in] Windows.UI.Xaml.Controls.MediaElement *pElement, [out, retval] Windows.Foundation.IAsyncOperation<IVisualizationSource *> **source);
-		HRESULT CreateSourceFromAudioNode([in] Windows.Media.Audio.IAudioNode *pNode, [out, retval] IVisualizationSource **source);
+		HRESULT CreateFromMediaElementAsync([in] Windows.UI.Xaml.Controls.MediaElement *pObject, [out, retval] Windows.Foundation.IAsyncOperation<IVisualizationSource *> **ppAsyncOp);
+		HRESULT CreateFromMediaPlayerAsync([in] Windows.Media.Playback.MediaPlayer *pObject, [out, retval] Windows.Foundation.IAsyncOperation<IVisualizationSource *> **ppAsyncOp);
+		HRESULT CreateFromAudioNode([in] Windows.Media.Audio.IAudioNode *pObject, [out, retval] IVisualizationSource **ppSource);
 	}
 
 	[uuid(B74989FE-1436-4179-92B6-E06B540E0C7E)]
