+++ AudioVisualizer/pch.h
@@ -1,5 +1,9 @@
 ﻿#pragma once
 
+#define _CRTDBG_MAP_ALLOC  
+#include <stdlib.h>  
+#include <crtdbg.h>
+
 #include "targetver.h"
 
 #ifndef WIN32_LEAN_AND_MEAN
