+++ AudioVisualizer/trace.h
@@ -1,6 +1,8 @@
 #pragma once
 #include <mfapi.h>
 #include <windows.foundation.diagnostics.h>
+#include "VisualizationData.h"
+#include "VisualizationDataFrame.h"
 
 namespace AudioVisualizer
 {
@@ -38,8 +40,8 @@ namespace AudioVisualizer
 			static HRESULT Log_SetInputPosition(long frameIndex);
 			static HRESULT Log_ProcessSample(IMFSample *pSample);
 			static HRESULT Log_StartCalculate(ABI::Windows::Foundation::Diagnostics::ILoggingActivity **ppActivity, long position, size_t bufferLength);
-			static HRESULT Log_GetData(REFERENCE_TIME currentPosition, IMFSample *pSample, IMFSample *pQueueFront, size_t queueSize, HRESULT result);
-			static HRESULT Log_OutputQueuePush(size_t queueSize);
+			static HRESULT Log_GetData(REFERENCE_TIME currentPosition, IVisualizationDataFrame *pFrame, IVisualizationDataFrame	*pQueueFront, size_t queueSize, HRESULT result);
+			static HRESULT Log_OutputQueuePush(IVisualizationDataFrame *pFrame,size_t queueSize);
 			static HRESULT Log_GetFromBuffer(REFERENCE_TIME position);
 			static HRESULT Log_StartCalculateFft(ABI::Windows::Foundation::Diagnostics::ILoggingActivity **ppActivity);
 			static HRESULT Log_LineNumber(UINT32 lineNumber);
